@@ -1,19 +1,19 @@
 name: dart_sdk_archive
 description: An absolute bare-bones web app.
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
-  http: ^0.12.0
-  path: ^1.6.2
-  quiver: ^2.0.0
-  platform_detect: ^1.3.5
+  http: ^0.13.1
+  path: ^1.8.0
+  quiver: ^3.0.1
+  platform_detect2: ^1.5.0+2
   sdk_builds:
     path: ../sdk_builds.dart
 
 dev_dependencies:
-  build_runner: ^1.1.2
-  build_web_compilers: ^2.0.0
-  dart_style: ^1.3.0
-  grinder: ^0.8.0
-  pedantic: ^1.0.0
-  test: ^1.14.0
+  build_runner: ^1.12.0
+  build_web_compilers: ^2.16.0
+  dart_style: ^2.0.0
+  grinder: ^0.9.0
+  pedantic: ^1.11.0
+  test: ^1.16.8
