@@ -1,15 +1,14 @@
 name: sdk_builds
 description: Utilities for accessing builds of the Dart SDK.
-version: 0.2.0-dev
-author: Kevin Moore <kevmoo@google.com>
-homepage: https://github.com/kevmoo/sdk_builds.dart
+version: 0.3.0
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
-  googleapis: ^0.53.0
-  http: ^0.12.0+2
-  path: ^1.6.2
-  pub_semver: ^1.4.2
-  quiver: ^2.0.3
+  googleapis: ^2.0.0
+  http: ^0.13.1
+  path: ^1.8.0
+  pub_semver: ^2.0.0
+  quiver: ^3.0.1
 dev_dependencies:
-  test: ^1.6.0
+  test: ^1.16.8
+  pedantic: ^1.11.0
