@@ -1,6 +1,5 @@
 import 'package:dart_sdk_archive/src/util.dart';
 import 'package:path/path.dart' as path;
-import 'package:quiver/async.dart' as quiver_async;
 import 'package:sdk_builds/sdk_builds.dart';
 
 class SvnVersionGenerator {
@@ -22,17 +21,18 @@ class SvnVersionGenerator {
 
   Future _loadVersionInfo(
       Map<String, VersionInfo> versionInfos, String channel) async {
-    var versionPaths = await _downloader.fetchVersionPaths(channel);
-    var versionBaseNames =
-        await versionPaths.map((s) => path.basename(s)).toList();
+    final versionBaseNames = await _downloader
+        .fetchVersionPaths(channel)
+        .map((s) => path.basename(s))
+        .toList();
 
-    await quiver_async.forEachAsync(versionBaseNames, (name) async {
+    await Future.forEach<String>(versionBaseNames, (name) async {
       if (!isSvnRevision(name)) {
         return;
       }
 
       var versionInfo = await _downloader.fetchVersion(channel, name);
       versionInfos[name] = versionInfo;
-    }, maxTasks: 6);
+    });
   }
 }
