@@ -1,152 +1,152 @@
 const Map<String, String> svnVersions = {
-  "29803": "0.8.10-rev.3.29803",
-  "30107": "0.8.10-rev.10.30107",
-  "30188": "1.0.0-rev.3.30188",
-  "31822": "1.1.1",
-  "30798": "1.0.0-rev.10.30798",
-  "30036": "0.8.10-rev.6.30036",
-  "32314": "1.1.3",
-  "33014": "1.2.0",
-  "34825": "1.3.0",
-  "35530": "1.3.6",
-  "36345": "1.4.0",
-  "35121": "1.3.3",
-  "36647": "1.4.2",
-  "38663": "1.5.8",
-  "37644": "1.5.1",
-  "37972": "1.5.3",
-  "37348": "1.4.3",
-  "37942": "1.5.2",
-  "39553": "1.6.0",
-  "42013": "1.8.0",
-  "41096": "1.7.2",
-  "42039": "1.8.3",
-  "42828": "1.8.5",
-  "44672": "1.9.1",
-  "45104": "1.9.3",
-  "45396": "1.10.0",
-  "45692": "1.10.1",
-  "30039": "0.8.10-rev.8.30039",
-  "29962": "0.8.10-rev.6.29962",
-  "30104": "0.8.10-rev.10.30104",
-  "30338": "1.0.0-rev.7.30338",
-  "30187": "1.0.0-rev.3.30187",
-  "30657": "1.0.1-rev.3.30657",
-  "30821": "1.0.2-rev.1.30821",
-  "31123": "1.1.0-dev.4.0",
-  "31329": "1.1.0-dev.5.0",
-  "30939": "1.0.3-rev.0.30939",
-  "31777": "1.1.0-dev.5.10",
-  "31661": "1.1.0-dev.5.6",
-  "31736": "1.1.0-dev.5.9",
-  "31918": "1.2.0-dev.1.0",
-  "31818": "1.1.0-dev.5.11",
-  "32164": "1.2.0-dev.2.4",
-  "32242": "1.2.0-dev.3.2",
-  "32426": "1.2.0-dev.4.0",
-  "32688": "1.2.0-dev.5.7",
-  "32712": "1.2.0-dev.5.8",
-  "32844": "1.2.0-dev.5.12",
-  "32778": "1.2.0-dev.5.11",
-  "32954": "1.2.0-dev.5.15",
-  "33060": "1.3.0-dev.0.0",
-  "33192": "1.3.0-dev.1.1",
-  "33495": "1.3.0-dev.3.2",
-  "34229": "1.3.0-dev.5.2",
-  "33731": "1.3.0-dev.4.1",
-  "34463": "1.3.0-dev.7.2",
-  "34284": "1.3.0-dev.6.1",
-  "34497": "1.3.0-dev.7.5",
-  "34591": "1.3.0-dev.7.7",
-  "34792": "1.3.0-dev.7.12",
-  "34756": "1.3.0-dev.7.11",
-  "35275": "1.4.0-dev.3.0",
-  "35068": "1.4.0-dev.2.2",
-  "34683": "1.3.0-dev.7.10",
-  "35677": "1.4.0-dev.5.1",
-  "35890": "1.4.0-dev.6.2",
-  "35960": "1.4.0-dev.6.3",
-  "36091": "1.4.0-dev.6.5",
-  "35362": "1.4.0-dev.4.0",
-  "36146": "1.4.0-dev.6.6",
-  "36210": "1.4.0-dev.6.7",
-  "36284": "1.4.0-dev.6.8",
-  "36412": "1.5.0-dev.0.0",
-  "36341": "1.4.0-dev.6.9",
-  "36630": "1.5.0-dev.2.0",
-  "36542": "1.5.0-dev.1.1",
-  "36871": "1.5.0-dev.3.4",
-  "37028": "1.5.0-dev.4.1",
-  "37071": "1.5.0-dev.4.2",
-  "37223": "1.5.0-dev.4.7",
-  "37161": "1.5.0-dev.4.5",
-  "37360": "1.5.0-dev.4.13",
-  "37251": "1.5.0-dev.4.8",
-  "37302": "1.5.0-dev.4.11",
-  "37385": "1.5.0-dev.4.14",
-  "37438": "1.5.0-dev.4.15",
-  "37532": "1.5.0-dev.4.17",
-  "36979": "1.5.0-dev.4.0",
-  "37580": "1.5.0-dev.4.20",
-  "37475": "1.5.0-dev.4.16",
-  "37639": "1.5.0-dev.4.23",
-  "37743": "1.6.0-dev.0.0",
-  "37846": "1.6.0-dev.0.1",
-  "37936": "1.6.0-dev.1.2",
-  "38083": "1.6.0-dev.2.0",
-  "38145": "1.6.0-dev.3.0",
-  "38380": "1.6.0-dev.4.0",
-  "38621": "1.6.0-dev.6.0",
-  "38831": "1.6.0-dev.7.0",
-  "38967": "1.6.0-dev.8.0",
-  "39285": "1.6.0-dev.9.3",
-  "39401": "1.6.0-dev.9.5",
-  "39442": "1.6.0-dev.9.6",
-  "39661": "1.7.0-dev.0.1",
-  "39537": "1.6.0-dev.9.7",
-  "40090": "1.7.0-dev.2.0",
-  "39799": "1.7.0-dev.1.0",
-  "40675": "1.7.0-dev.4.0",
-  "40302": "1.7.0-dev.3.0",
-  "40806": "1.7.0-dev.4.1",
-  "40917": "1.7.0-dev.4.3",
-  "40987": "1.7.0-dev.4.4",
-  "41004": "1.7.0-dev.4.5",
-  "41090": "1.7.0-dev.4.6",
-  "41275": "1.8.0-dev.1.1",
-  "41389": "1.8.0-dev.2.0",
-  "41515": "1.8.0-dev.3.0",
-  "41684": "1.8.0-dev.4.0",
-  "41762": "1.8.0-dev.4.1",
-  "41923": "1.8.0-dev.4.5",
-  "41847": "1.8.0-dev.4.4",
-  "41793": "1.8.0-dev.4.2",
-  "41978": "1.8.0-dev.4.6",
-  "42033": "1.9.0-dev.0.0",
-  "41145": "1.8.0-dev.0.0",
-  "42684": "1.9.0-dev.3.0",
-  "42546": "1.9.0-dev.2.2",
-  "42856": "1.9.0-dev.4.0",
-  "42241": "1.9.0-dev.1.0",
-  "43384": "1.9.0-dev.5.1",
-  "43584": "1.9.0-dev.7.1",
-  "43903": "1.9.0-dev.8.4",
-  "44224": "1.9.0-dev.10.0",
-  "43715": "1.9.0-dev.8.0",
-  "44018": "1.9.0-dev.9.1",
-  "44260": "1.9.0-dev.10.2",
-  "44314": "1.9.0-dev.10.4",
-  "44550": "1.9.0-dev.10.10",
-  "44500": "1.9.0-dev.10.7",
-  "44532": "1.9.0-dev.10.9",
-  "44630": "1.9.0-dev.10.13",
-  "44728": "1.10.0-dev.0.1",
-  "44601": "1.9.0-dev.10.12",
-  "45054": "1.10.0-dev.1.0",
-  "45089": "1.10.0-dev.1.1",
-  "45201": "1.10.0-dev.1.5",
-  "45268": "1.10.0-dev.1.7",
-  "45369": "1.10.0-dev.1.10",
-  "45311": "1.10.0-dev.1.9",
-  "45519": "1.11.0-dev.0.0"
+  '29803': '0.8.10-rev.3.29803',
+  '30107': '0.8.10-rev.10.30107',
+  '30188': '1.0.0-rev.3.30188',
+  '31822': '1.1.1',
+  '30798': '1.0.0-rev.10.30798',
+  '30036': '0.8.10-rev.6.30036',
+  '32314': '1.1.3',
+  '33014': '1.2.0',
+  '34825': '1.3.0',
+  '35530': '1.3.6',
+  '36345': '1.4.0',
+  '35121': '1.3.3',
+  '36647': '1.4.2',
+  '38663': '1.5.8',
+  '37644': '1.5.1',
+  '37972': '1.5.3',
+  '37348': '1.4.3',
+  '37942': '1.5.2',
+  '39553': '1.6.0',
+  '42013': '1.8.0',
+  '41096': '1.7.2',
+  '42039': '1.8.3',
+  '42828': '1.8.5',
+  '44672': '1.9.1',
+  '45104': '1.9.3',
+  '45396': '1.10.0',
+  '45692': '1.10.1',
+  '30039': '0.8.10-rev.8.30039',
+  '29962': '0.8.10-rev.6.29962',
+  '30104': '0.8.10-rev.10.30104',
+  '30338': '1.0.0-rev.7.30338',
+  '30187': '1.0.0-rev.3.30187',
+  '30657': '1.0.1-rev.3.30657',
+  '30821': '1.0.2-rev.1.30821',
+  '31123': '1.1.0-dev.4.0',
+  '31329': '1.1.0-dev.5.0',
+  '30939': '1.0.3-rev.0.30939',
+  '31777': '1.1.0-dev.5.10',
+  '31661': '1.1.0-dev.5.6',
+  '31736': '1.1.0-dev.5.9',
+  '31918': '1.2.0-dev.1.0',
+  '31818': '1.1.0-dev.5.11',
+  '32164': '1.2.0-dev.2.4',
+  '32242': '1.2.0-dev.3.2',
+  '32426': '1.2.0-dev.4.0',
+  '32688': '1.2.0-dev.5.7',
+  '32712': '1.2.0-dev.5.8',
+  '32844': '1.2.0-dev.5.12',
+  '32778': '1.2.0-dev.5.11',
+  '32954': '1.2.0-dev.5.15',
+  '33060': '1.3.0-dev.0.0',
+  '33192': '1.3.0-dev.1.1',
+  '33495': '1.3.0-dev.3.2',
+  '34229': '1.3.0-dev.5.2',
+  '33731': '1.3.0-dev.4.1',
+  '34463': '1.3.0-dev.7.2',
+  '34284': '1.3.0-dev.6.1',
+  '34497': '1.3.0-dev.7.5',
+  '34591': '1.3.0-dev.7.7',
+  '34792': '1.3.0-dev.7.12',
+  '34756': '1.3.0-dev.7.11',
+  '35275': '1.4.0-dev.3.0',
+  '35068': '1.4.0-dev.2.2',
+  '34683': '1.3.0-dev.7.10',
+  '35677': '1.4.0-dev.5.1',
+  '35890': '1.4.0-dev.6.2',
+  '35960': '1.4.0-dev.6.3',
+  '36091': '1.4.0-dev.6.5',
+  '35362': '1.4.0-dev.4.0',
+  '36146': '1.4.0-dev.6.6',
+  '36210': '1.4.0-dev.6.7',
+  '36284': '1.4.0-dev.6.8',
+  '36412': '1.5.0-dev.0.0',
+  '36341': '1.4.0-dev.6.9',
+  '36630': '1.5.0-dev.2.0',
+  '36542': '1.5.0-dev.1.1',
+  '36871': '1.5.0-dev.3.4',
+  '37028': '1.5.0-dev.4.1',
+  '37071': '1.5.0-dev.4.2',
+  '37223': '1.5.0-dev.4.7',
+  '37161': '1.5.0-dev.4.5',
+  '37360': '1.5.0-dev.4.13',
+  '37251': '1.5.0-dev.4.8',
+  '37302': '1.5.0-dev.4.11',
+  '37385': '1.5.0-dev.4.14',
+  '37438': '1.5.0-dev.4.15',
+  '37532': '1.5.0-dev.4.17',
+  '36979': '1.5.0-dev.4.0',
+  '37580': '1.5.0-dev.4.20',
+  '37475': '1.5.0-dev.4.16',
+  '37639': '1.5.0-dev.4.23',
+  '37743': '1.6.0-dev.0.0',
+  '37846': '1.6.0-dev.0.1',
+  '37936': '1.6.0-dev.1.2',
+  '38083': '1.6.0-dev.2.0',
+  '38145': '1.6.0-dev.3.0',
+  '38380': '1.6.0-dev.4.0',
+  '38621': '1.6.0-dev.6.0',
+  '38831': '1.6.0-dev.7.0',
+  '38967': '1.6.0-dev.8.0',
+  '39285': '1.6.0-dev.9.3',
+  '39401': '1.6.0-dev.9.5',
+  '39442': '1.6.0-dev.9.6',
+  '39661': '1.7.0-dev.0.1',
+  '39537': '1.6.0-dev.9.7',
+  '40090': '1.7.0-dev.2.0',
+  '39799': '1.7.0-dev.1.0',
+  '40675': '1.7.0-dev.4.0',
+  '40302': '1.7.0-dev.3.0',
+  '40806': '1.7.0-dev.4.1',
+  '40917': '1.7.0-dev.4.3',
+  '40987': '1.7.0-dev.4.4',
+  '41004': '1.7.0-dev.4.5',
+  '41090': '1.7.0-dev.4.6',
+  '41275': '1.8.0-dev.1.1',
+  '41389': '1.8.0-dev.2.0',
+  '41515': '1.8.0-dev.3.0',
+  '41684': '1.8.0-dev.4.0',
+  '41762': '1.8.0-dev.4.1',
+  '41923': '1.8.0-dev.4.5',
+  '41847': '1.8.0-dev.4.4',
+  '41793': '1.8.0-dev.4.2',
+  '41978': '1.8.0-dev.4.6',
+  '42033': '1.9.0-dev.0.0',
+  '41145': '1.8.0-dev.0.0',
+  '42684': '1.9.0-dev.3.0',
+  '42546': '1.9.0-dev.2.2',
+  '42856': '1.9.0-dev.4.0',
+  '42241': '1.9.0-dev.1.0',
+  '43384': '1.9.0-dev.5.1',
+  '43584': '1.9.0-dev.7.1',
+  '43903': '1.9.0-dev.8.4',
+  '44224': '1.9.0-dev.10.0',
+  '43715': '1.9.0-dev.8.0',
+  '44018': '1.9.0-dev.9.1',
+  '44260': '1.9.0-dev.10.2',
+  '44314': '1.9.0-dev.10.4',
+  '44550': '1.9.0-dev.10.10',
+  '44500': '1.9.0-dev.10.7',
+  '44532': '1.9.0-dev.10.9',
+  '44630': '1.9.0-dev.10.13',
+  '44728': '1.10.0-dev.0.1',
+  '44601': '1.9.0-dev.10.12',
+  '45054': '1.10.0-dev.1.0',
+  '45089': '1.10.0-dev.1.1',
+  '45201': '1.10.0-dev.1.5',
+  '45268': '1.10.0-dev.1.7',
+  '45369': '1.10.0-dev.1.10',
+  '45311': '1.10.0-dev.1.9',
+  '45519': '1.11.0-dev.0.0'
 };
