@@ -5,11 +5,11 @@ import 'package:dart_sdk_archive/src/generator.dart';
 import 'package:grinder/grinder.dart';
 import 'package:dart_style/dart_style.dart';
 
-main(args) => grind(args);
+void main(args) => grind(args);
 
 @DefaultTask('Build SVN Revision JSON')
-buildJson() async {
-  print("fetching SVN versions...");
+void buildJson() async {
+  print('fetching SVN versions...');
   var filePath = 'lib/src/svn_versions.dart';
   var generator = SvnVersionGenerator();
   var svnVersions = await generator.svnVersions;
